Renames TextManager to TranslationManager
