+++ src/main/java/fr/openmc/core/commands/fun/Playtime.java
@@ -17,7 +17,7 @@ public class Playtime {
     @Description("Donne votre temps de jeu")
     private void playtime(Player player) {
         long timePlayed = player.getStatistic(Statistic.PLAY_ONE_MINUTE);
-        MessagesManager.sendMessageType(player, OMCPlugin.getTextManager().getText("messages.general.playtime",
+        MessagesManager.sendMessageType(player, OMCPlugin.getTranslationManager().getTranslation("messages.general.playtime",
                         "fr", "playTime", DateUtils.convertTime(timePlayed)),
                 Prefix.OPENMC, MessageType.INFO, true);
     }
